% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{create.api}
\alias{create.api}
\title{Obtain API key.}
\usage{
create.api(auth.token)
}
\arguments{
\item{auth.token}{character. Authorization token from Global Forest Watch (see description)}
}
\value{
Returns an API key from Global Forest Watch.
}
\description{
Request an API key for downloading data from Global Forest Watch. Requires an authorization token which can be obtained using the function \code{\link{request.token}}.
}
\examples{
\dontrun{
  # Example to obtain an API key
  auth_token <- request.token(username = "your_username", pw = "your_password")
  api_key <- create.api(auth_token)
}
}
