% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{find.dataset}
\alias{find.dataset}
\title{Print Global Forest Watch Data.}
\usage{
find.dataset(api.key)
}
\arguments{
\item{api.key}{character. API key from Global Forest Watch (see description)}
}
\value{
NULL. Prints the names of all available datasets from Global Forest Watch.
}
\description{
Provides the names of all the datasets available from Global Forest Watch. Requires an API key from \code{\link{create.api}}.
}
