% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{find.target.MY}
\alias{find.target.MY}
\title{Define target month based on current date.}
\usage{
find.target.MY(current_date, months_back)
}
\arguments{
\item{current_date}{date. Date object representing the current date, typically obtained using `Sys.Date()`.}

\item{months_back}{integer. A positive integer representing the number of months to move backwards from the current date to find the target month. For example, the month prior to the current month is 1 back.}
}
\value{
character vector of length 2. The output from \code{\link{find.target.MY}}, as c("MM", "YYYY"), where the first element is the target month ("MM") and the second element is the target year ("YYYY").
}
\description{
Calculate the target month and year for downloading deforestation alerts based on the current date and a specified number of months back. This is useful for time-based data queries to Global Forest Watch.
}
