% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{geom.define}
\alias{geom.define}
\title{Define extent.}
\usage{
geom.define(xmn, xmx, ymn, ymx)
}
\arguments{
\item{xmn}{numeric. xmin for bounding box.}

\item{xmx}{numeric. xmax for bounding box.}

\item{ymn}{numeric. ymin for bounding box.}

\item{ymx}{numeric. ymax for bounding box.}
}
\value{
Returns a list with bounding coordinates.
}
\description{
Define a geographical extent for requesting data from Global Forest Watch.
}
