% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{get.gfw.alerts}
\alias{get.gfw.alerts}
\title{Download Global Forest Watch Integrated Alerts.}
\usage{
get.gfw.alerts(
  api.key,
  geom.ext,
  start,
  end,
  confidence = "highest",
  max_retries = 3,
  retry_delay = 5
)
}
\arguments{
\item{api.key}{character. API key from Global Forest Watch (see description)}

\item{geom.ext}{numeric vector of length 4. Boundary coordinates for the geographical extent, given as (xmin, xmax, ymin, ymax).}

\item{start}{character string. The start date for the alerts (in "YYYY-MM-DD" format).}

\item{end}{character string. The end date for the alerts (in "YYYY-MM-DD" format).}

\item{confidence}{optional character. Define the confidence level for alerts (high or highest, default is highest)}

\item{max_retries}{optional single positive numeric value. Maximum number of times to resend the request if it fails (default is 3 tries).}

\item{retry_delay}{optional single positive numeric value. Number of seconds delay between retries so as to not overload the request system (default is 5 seconds).}
}
\value{
Returns the Global Forest Watch integrated deforestation alerts as a point SpatVector object. If the download fails after retries, returns `NULL`.
}
\description{
Download Global Forest Watch Integrated Alerts for deforestation across a defined geographical extent and time period. Requires an API key from \code{\link{create.api}} and a vector of boundary coordinates for the geographical extent.
}
