% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{request.token}
\alias{request.token}
\title{Request token.}
\usage{
request.token(username, pw)
}
\arguments{
\item{username}{character. The username for your Global Forest Watch account.}

\item{pw}{character. The password for your Global Forest Watch account.}
}
\value{
Returns an authorization token from Global Forest Watch.
}
\description{
Request an authorization token from GFW. Requires your username and password.
}
\examples{
\dontrun{
  # Example to request a token.
  auth_token <- request.token(username = "your_username", pw = "your_password")
}
}
