% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{run.download.month}
\alias{run.download.month}
\title{Monthly Global Forest Watch Integrated Alerts.}
\usage{
run.download.month(
  target,
  api.key,
  geom.ext,
  target.proj = "EPSG:4326",
  confidence = "highest",
  max_retries = 3,
  retry_delay = 5
)
}
\arguments{
\item{target}{character vector of length 2. The output from \code{\link{find.target.MY}}, where the first element is the target month ("MM") and the second element is the target year ("YYYY").}

\item{api.key}{character. API key from Global Forest Watch (see description)}

\item{geom.ext}{numeric vector of length 4. Boundary coordinates for the geographical extent (xmin, xmax, ymin, ymax). See \code{\link{geom.define}} for more details.}

\item{target.proj}{character. The EPSG code for the target projection of the spatial data (e.g., "EPSG:4326").}

\item{confidence}{optional character. Define the confidence level for alerts (high or highest, default is highest)}

\item{max_retries}{optional single positive numeric value. Maximum number of times to resend the request if it fails (default is 3 tries).}

\item{retry_delay}{optional single positive numeric value. Number of seconds delay between retries so as to not overload the request system (default is 5 seconds).}
}
\value{
A list with a SpatVector point object consisting of the Global Forest Watch integrated deforestation alerts for the target month, plus a start and end date.
}
\description{
Download monthly Global Forest Watch Integrated Alerts for deforestation across a defined geographical extent. The function splits the target month into short time steps to minimize the risk of the request failing and combines the results. For requests that don't require results as monthly time steps, use \code{\link{get.gfw.alerts}}.
}
