% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{show.fields}
\alias{show.fields}
\title{Print data fields.}
\usage{
show.fields(url)
}
\arguments{
\item{url}{character. url for the dataset from Global Forest Watch (see description)}
}
\value{
Returns a list of fields associated with the dataset defined by your url.
}
\description{
Print all fields associated with a Global Forest Watch dataset. Requires the data url which can be obtained from \code{\link{find.dataset}}.
}
