% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadtools.R
\name{split.month}
\alias{split.month}
\title{Split a month into small time steps.}
\usage{
\method{split}{month}(target)
}
\arguments{
\item{target}{character vector of length 2. The output from \code{\link{find.target.MY}}, where the first element is the target month ("MM") and the second element is the target year ("YYYY").}
}
\value{
A list consisting of a sequence of start and end dates to split the month, plus the first and last dates in the month.
}
\description{
Splits a target month into steps of 2 days.
}
